requirements.txt
peewee
requests
Adafruit_Python_DHT
RPi.GPIO
PyTelegramBotAPI==2.2.3
python-telegram-bot
